requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.4
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
psycopg2-binary==2.9.10 